+++ gen/com/kaylerrenslow/armaDialogCreator/expression/ExpressionParser.java
@@ -19,10 +19,10 @@ public class ExpressionParser extends Parser {
 	public static final int
 		String=1, Quote=2, DQuote=3, LCurly=4, RCurly=5, LBracket=6, RBracket=7, 
 		Plus=8, Minus=9, FSlash=10, Star=11, LParen=12, RParen=13, Comma=14, Min=15, 
-		Max=16, If=17, Then=18, Else=19, ExitWith=20, Select=21, Equal=22, Semicolon=23, 
-		Identifier=24, IntegerLiteral=25, FloatLiteral=26, Digits=27, DecSignificand=28, 
-		DecExponent=29, HexLiteral=30, HexDigits=31, Letter=32, LetterOrDigit=33, 
-		WhiteSpace=34;
+		Max=16, If=17, Then=18, Else=19, ExitWith=20, Select=21, EqEq=22, NotEq=23, 
+		Lt=24, LtEq=25, Gt=26, GtEq=27, Equal=28, Semicolon=29, Identifier=30, 
+		IntegerLiteral=31, FloatLiteral=32, Digits=33, DecSignificand=34, DecExponent=35, 
+		HexLiteral=36, HexDigits=37, Letter=38, LetterOrDigit=39, WhiteSpace=40;
 	public static final int
 		RULE_statements = 0, RULE_statement = 1, RULE_assignment = 2, RULE_code = 3, 
 		RULE_expression = 4, RULE_unary_expression = 5, RULE_paren_expression = 6, 
@@ -37,14 +37,16 @@ public class ExpressionParser extends Parser {
 	private static final String[] _LITERAL_NAMES = {
 		null, null, "'''", "'\"'", "'{'", "'}'", "'['", "']'", "'+'", "'-'", "'/'", 
 		"'*'", "'('", "')'", "','", "'min'", "'max'", "'if'", "'then'", "'else'", 
-		"'exitWith'", "'select'", "'='", "';'"
+		"'exitWith'", "'select'", "'=='", "'!='", "'<'", "'<='", "'>'", "'>='", 
+		"'='", "';'"
 	};
 	private static final String[] _SYMBOLIC_NAMES = {
 		null, "String", "Quote", "DQuote", "LCurly", "RCurly", "LBracket", "RBracket", 
 		"Plus", "Minus", "FSlash", "Star", "LParen", "RParen", "Comma", "Min", 
-		"Max", "If", "Then", "Else", "ExitWith", "Select", "Equal", "Semicolon", 
-		"Identifier", "IntegerLiteral", "FloatLiteral", "Digits", "DecSignificand", 
-		"DecExponent", "HexLiteral", "HexDigits", "Letter", "LetterOrDigit", "WhiteSpace"
+		"Max", "If", "Then", "Else", "ExitWith", "Select", "EqEq", "NotEq", "Lt", 
+		"LtEq", "Gt", "GtEq", "Equal", "Semicolon", "Identifier", "IntegerLiteral", 
+		"FloatLiteral", "Digits", "DecSignificand", "DecExponent", "HexLiteral", 
+		"HexDigits", "Letter", "LetterOrDigit", "WhiteSpace"
 	};
 	public static final Vocabulary VOCABULARY = new VocabularyImpl(_LITERAL_NAMES, _SYMBOLIC_NAMES);
 
@@ -352,6 +354,7 @@ public class ExpressionParser extends Parser {
 		public ExpressionContext lf;
 		public ExpressionContext la;
 		public ExpressionContext lm;
+		public ExpressionContext lcomp;
 		public ExpressionContext lmax;
 		public ExpressionContext lmin;
 		public ExpressionContext select_e;
@@ -364,6 +367,8 @@ public class ExpressionParser extends Parser {
 		public ExpressionContext rf;
 		public ExpressionContext ra;
 		public ExpressionContext rm;
+		public Token compOp;
+		public ExpressionContext rcomp;
 		public ExpressionContext rmax;
 		public ExpressionContext rmin;
 		public ExpressionContext select_i;
@@ -392,6 +397,12 @@ public class ExpressionParser extends Parser {
 		public TerminalNode FSlash() { return getToken(ExpressionParser.FSlash, 0); }
 		public TerminalNode Plus() { return getToken(ExpressionParser.Plus, 0); }
 		public TerminalNode Minus() { return getToken(ExpressionParser.Minus, 0); }
+		public TerminalNode EqEq() { return getToken(ExpressionParser.EqEq, 0); }
+		public TerminalNode NotEq() { return getToken(ExpressionParser.NotEq, 0); }
+		public TerminalNode LtEq() { return getToken(ExpressionParser.LtEq, 0); }
+		public TerminalNode Lt() { return getToken(ExpressionParser.Lt, 0); }
+		public TerminalNode GtEq() { return getToken(ExpressionParser.GtEq, 0); }
+		public TerminalNode Gt() { return getToken(ExpressionParser.Gt, 0); }
 		public TerminalNode Max() { return getToken(ExpressionParser.Max, 0); }
 		public TerminalNode Min() { return getToken(ExpressionParser.Min, 0); }
 		public TerminalNode Select() { return getToken(ExpressionParser.Select, 0); }
@@ -420,6 +431,7 @@ public class ExpressionParser extends Parser {
 		ExpressionContext _prevctx = _localctx;
 		int _startState = 8;
 		enterRecursionRule(_localctx, 8, RULE_expression, _p);
+		int _la;
 		try {
 			int _alt;
 			enterOuterAlt(_localctx, 1);
@@ -472,7 +484,7 @@ public class ExpressionParser extends Parser {
 				throw new NoViableAltException(this);
 			}
 			_ctx.stop = _input.LT(-1);
-			setState(116);
+			setState(121);
 			_errHandler.sync(this);
 			_alt = getInterpreter().adaptivePredict(_input,6,_ctx);
 			while ( _alt!=2 && _alt!=org.antlr.v4.runtime.atn.ATN.INVALID_ALT_NUMBER ) {
@@ -480,7 +492,7 @@ public class ExpressionParser extends Parser {
 					if ( _parseListeners!=null ) triggerExitRuleEvent();
 					_prevctx = _localctx;
 					{
-					setState(114);
+					setState(119);
 					_errHandler.sync(this);
 					switch ( getInterpreter().adaptivePredict(_input,5,_ctx) ) {
 					case 1:
@@ -490,11 +502,11 @@ public class ExpressionParser extends Parser {
 						_localctx.ls = _prevctx;
 						pushNewRecursionContext(_localctx, _startState, RULE_expression);
 						setState(79);
-						if (!(precpred(_ctx, 10))) throw new FailedPredicateException(this, "precpred(_ctx, 10)");
+						if (!(precpred(_ctx, 11))) throw new FailedPredicateException(this, "precpred(_ctx, 11)");
 						setState(80);
 						match(Star);
 						setState(81);
-						((ExpressionContext)_localctx).rs = expression(11);
+						((ExpressionContext)_localctx).rs = expression(12);
 						((ExpressionContext)_localctx).ast =  new AST.MultExpr(((ExpressionContext)_localctx).ls.ast, ((ExpressionContext)_localctx).rs.ast);
 						}
 						break;
@@ -505,11 +517,11 @@ public class ExpressionParser extends Parser {
 						_localctx.lf = _prevctx;
 						pushNewRecursionContext(_localctx, _startState, RULE_expression);
 						setState(84);
-						if (!(precpred(_ctx, 9))) throw new FailedPredicateException(this, "precpred(_ctx, 9)");
+						if (!(precpred(_ctx, 10))) throw new FailedPredicateException(this, "precpred(_ctx, 10)");
 						setState(85);
 						match(FSlash);
 						setState(86);
-						((ExpressionContext)_localctx).rf = expression(10);
+						((ExpressionContext)_localctx).rf = expression(11);
 						((ExpressionContext)_localctx).ast =  new AST.DivExpr(((ExpressionContext)_localctx).lf.ast, ((ExpressionContext)_localctx).rf.ast);
 						}
 						break;
@@ -520,11 +532,11 @@ public class ExpressionParser extends Parser {
 						_localctx.la = _prevctx;
 						pushNewRecursionContext(_localctx, _startState, RULE_expression);
 						setState(89);
-						if (!(precpred(_ctx, 8))) throw new FailedPredicateException(this, "precpred(_ctx, 8)");
+						if (!(precpred(_ctx, 9))) throw new FailedPredicateException(this, "precpred(_ctx, 9)");
 						setState(90);
 						match(Plus);
 						setState(91);
-						((ExpressionContext)_localctx).ra = expression(9);
+						((ExpressionContext)_localctx).ra = expression(10);
 						((ExpressionContext)_localctx).ast =  new AST.AddExpr(((ExpressionContext)_localctx).la.ast, ((ExpressionContext)_localctx).ra.ast);
 						}
 						break;
@@ -535,55 +547,79 @@ public class ExpressionParser extends Parser {
 						_localctx.lm = _prevctx;
 						pushNewRecursionContext(_localctx, _startState, RULE_expression);
 						setState(94);
-						if (!(precpred(_ctx, 7))) throw new FailedPredicateException(this, "precpred(_ctx, 7)");
+						if (!(precpred(_ctx, 8))) throw new FailedPredicateException(this, "precpred(_ctx, 8)");
 						setState(95);
 						match(Minus);
 						setState(96);
-						((ExpressionContext)_localctx).rm = expression(8);
+						((ExpressionContext)_localctx).rm = expression(9);
 						((ExpressionContext)_localctx).ast =  new AST.SubExpr(((ExpressionContext)_localctx).lm.ast, ((ExpressionContext)_localctx).rm.ast);
 						}
 						break;
 					case 5:
 						{
 						_localctx = new ExpressionContext(_parentctx, _parentState);
-						_localctx.lmax = _prevctx;
-						_localctx.lmax = _prevctx;
+						_localctx.lcomp = _prevctx;
+						_localctx.lcomp = _prevctx;
 						pushNewRecursionContext(_localctx, _startState, RULE_expression);
 						setState(99);
-						if (!(precpred(_ctx, 5))) throw new FailedPredicateException(this, "precpred(_ctx, 5)");
+						if (!(precpred(_ctx, 7))) throw new FailedPredicateException(this, "precpred(_ctx, 7)");
 						setState(100);
-						match(Max);
+						((ExpressionContext)_localctx).compOp = _input.LT(1);
+						_la = _input.LA(1);
+						if ( !((((_la) & ~0x3f) == 0 && ((1L << _la) & ((1L << EqEq) | (1L << NotEq) | (1L << Lt) | (1L << LtEq) | (1L << Gt) | (1L << GtEq))) != 0)) ) {
+							((ExpressionContext)_localctx).compOp = (Token)_errHandler.recoverInline(this);
+						}
+						else {
+							if ( _input.LA(1)==Token.EOF ) matchedEOF = true;
+							_errHandler.reportMatch(this);
+							consume();
+						}
 						setState(101);
-						((ExpressionContext)_localctx).rmax = expression(6);
-						((ExpressionContext)_localctx).ast =  new AST.MaxExpr(((ExpressionContext)_localctx).lmax.ast, ((ExpressionContext)_localctx).rmax.ast);
+						((ExpressionContext)_localctx).rcomp = expression(8);
+						((ExpressionContext)_localctx).ast =  new AST.CompExpr(((ExpressionContext)_localctx).lcomp.ast, ((ExpressionContext)_localctx).rcomp.ast, (((ExpressionContext)_localctx).compOp!=null?((ExpressionContext)_localctx).compOp.getText():null));
 						}
 						break;
 					case 6:
 						{
 						_localctx = new ExpressionContext(_parentctx, _parentState);
-						_localctx.lmin = _prevctx;
-						_localctx.lmin = _prevctx;
+						_localctx.lmax = _prevctx;
+						_localctx.lmax = _prevctx;
 						pushNewRecursionContext(_localctx, _startState, RULE_expression);
 						setState(104);
-						if (!(precpred(_ctx, 4))) throw new FailedPredicateException(this, "precpred(_ctx, 4)");
+						if (!(precpred(_ctx, 5))) throw new FailedPredicateException(this, "precpred(_ctx, 5)");
 						setState(105);
-						match(Min);
+						match(Max);
 						setState(106);
-						((ExpressionContext)_localctx).rmin = expression(5);
-						((ExpressionContext)_localctx).ast =  new AST.MinExpr(((ExpressionContext)_localctx).lmin.ast, ((ExpressionContext)_localctx).rmin.ast);
+						((ExpressionContext)_localctx).rmax = expression(6);
+						((ExpressionContext)_localctx).ast =  new AST.MaxExpr(((ExpressionContext)_localctx).lmax.ast, ((ExpressionContext)_localctx).rmax.ast);
 						}
 						break;
 					case 7:
 						{
 						_localctx = new ExpressionContext(_parentctx, _parentState);
+						_localctx.lmin = _prevctx;
+						_localctx.lmin = _prevctx;
+						pushNewRecursionContext(_localctx, _startState, RULE_expression);
+						setState(109);
+						if (!(precpred(_ctx, 4))) throw new FailedPredicateException(this, "precpred(_ctx, 4)");
+						setState(110);
+						match(Min);
+						setState(111);
+						((ExpressionContext)_localctx).rmin = expression(5);
+						((ExpressionContext)_localctx).ast =  new AST.MinExpr(((ExpressionContext)_localctx).lmin.ast, ((ExpressionContext)_localctx).rmin.ast);
+						}
+						break;
+					case 8:
+						{
+						_localctx = new ExpressionContext(_parentctx, _parentState);
 						_localctx.select_e = _prevctx;
 						_localctx.select_e = _prevctx;
 						pushNewRecursionContext(_localctx, _startState, RULE_expression);
-						setState(109);
+						setState(114);
 						if (!(precpred(_ctx, 3))) throw new FailedPredicateException(this, "precpred(_ctx, 3)");
-						setState(110);
+						setState(115);
 						match(Select);
-						setState(111);
+						setState(116);
 						((ExpressionContext)_localctx).select_i = expression(4);
 						((ExpressionContext)_localctx).ast =  new AST.SelectExpr(((ExpressionContext)_localctx).select_e.ast, ((ExpressionContext)_localctx).select_i.ast);
 						}
@@ -591,7 +627,7 @@ public class ExpressionParser extends Parser {
 					}
 					} 
 				}
-				setState(118);
+				setState(123);
 				_errHandler.sync(this);
 				_alt = getInterpreter().adaptivePredict(_input,6,_ctx);
 			}
@@ -640,15 +676,15 @@ public class ExpressionParser extends Parser {
 		Unary_expressionContext _localctx = new Unary_expressionContext(_ctx, getState());
 		enterRule(_localctx, 10, RULE_unary_expression);
 		try {
-			setState(135);
+			setState(140);
 			_errHandler.sync(this);
 			switch ( getInterpreter().adaptivePredict(_input,7,_ctx) ) {
 			case 1:
 				enterOuterAlt(_localctx, 1);
 				{
-				setState(119);
+				setState(124);
 				match(Plus);
-				setState(120);
+				setState(125);
 				((Unary_expressionContext)_localctx).ep = paren_expression();
 				((Unary_expressionContext)_localctx).ast =  new AST.UnaryExpr(true, ((Unary_expressionContext)_localctx).ep.ast);
 				}
@@ -656,9 +692,9 @@ public class ExpressionParser extends Parser {
 			case 2:
 				enterOuterAlt(_localctx, 2);
 				{
-				setState(123);
+				setState(128);
 				match(Plus);
-				setState(124);
+				setState(129);
 				((Unary_expressionContext)_localctx).ep1 = literal_expression();
 				((Unary_expressionContext)_localctx).ast =  new AST.UnaryExpr(true, ((Unary_expressionContext)_localctx).ep1.ast);
 				}
@@ -666,9 +702,9 @@ public class ExpressionParser extends Parser {
 			case 3:
 				enterOuterAlt(_localctx, 3);
 				{
-				setState(127);
+				setState(132);
 				match(Minus);
-				setState(128);
+				setState(133);
 				((Unary_expressionContext)_localctx).em = paren_expression();
 				((Unary_expressionContext)_localctx).ast =  new AST.UnaryExpr(false, ((Unary_expressionContext)_localctx).em.ast);
 				}
@@ -676,9 +712,9 @@ public class ExpressionParser extends Parser {
 			case 4:
 				enterOuterAlt(_localctx, 4);
 				{
-				setState(131);
+				setState(136);
 				match(Minus);
-				setState(132);
+				setState(137);
 				((Unary_expressionContext)_localctx).em1 = literal_expression();
 				((Unary_expressionContext)_localctx).ast =  new AST.UnaryExpr(false, ((Unary_expressionContext)_localctx).em1.ast);
 				}
@@ -724,11 +760,11 @@ public class ExpressionParser extends Parser {
 		try {
 			enterOuterAlt(_localctx, 1);
 			{
-			setState(137);
+			setState(142);
 			match(LParen);
-			setState(138);
+			setState(143);
 			((Paren_expressionContext)_localctx).e = expression(0);
-			setState(139);
+			setState(144);
 			match(RParen);
 			((Paren_expressionContext)_localctx).ast =  new AST.ParenExpr(((Paren_expressionContext)_localctx).e.ast);
 			}
@@ -780,13 +816,13 @@ public class ExpressionParser extends Parser {
 		Literal_expressionContext _localctx = new Literal_expressionContext(_ctx, getState());
 		enterRule(_localctx, 14, RULE_literal_expression);
 		try {
-			setState(155);
+			setState(160);
 			_errHandler.sync(this);
 			switch (_input.LA(1)) {
 			case Identifier:
 				enterOuterAlt(_localctx, 1);
 				{
-				setState(142);
+				setState(147);
 				((Literal_expressionContext)_localctx).id = match(Identifier);
 				((Literal_expressionContext)_localctx).ast =  new AST.IdentifierExpr((((Literal_expressionContext)_localctx).id!=null?((Literal_expressionContext)_localctx).id.getText():null));
 				}
@@ -795,7 +831,7 @@ public class ExpressionParser extends Parser {
 			case HexLiteral:
 				enterOuterAlt(_localctx, 2);
 				{
-				setState(144);
+				setState(149);
 				((Literal_expressionContext)_localctx).i = int_value();
 				((Literal_expressionContext)_localctx).ast =  new AST.IntegerExpr(((Literal_expressionContext)_localctx).i.i);
 				}
@@ -803,7 +839,7 @@ public class ExpressionParser extends Parser {
 			case FloatLiteral:
 				enterOuterAlt(_localctx, 3);
 				{
-				setState(147);
+				setState(152);
 				((Literal_expressionContext)_localctx).f = float_value();
 				((Literal_expressionContext)_localctx).ast =  new AST.FloatExpr(((Literal_expressionContext)_localctx).f.d);
 				}
@@ -811,7 +847,7 @@ public class ExpressionParser extends Parser {
 			case String:
 				enterOuterAlt(_localctx, 4);
 				{
-				setState(150);
+				setState(155);
 				((Literal_expressionContext)_localctx).s = match(String);
 				((Literal_expressionContext)_localctx).ast =  new AST.StringExpr((((Literal_expressionContext)_localctx).s!=null?((Literal_expressionContext)_localctx).s.getText():null));
 				}
@@ -819,7 +855,7 @@ public class ExpressionParser extends Parser {
 			case LBracket:
 				enterOuterAlt(_localctx, 5);
 				{
-				setState(152);
+				setState(157);
 				((Literal_expressionContext)_localctx).a = array();
 				((Literal_expressionContext)_localctx).ast =  ((Literal_expressionContext)_localctx).a.ast;
 				}
@@ -879,19 +915,19 @@ public class ExpressionParser extends Parser {
 		try {
 			enterOuterAlt(_localctx, 1);
 			{
-			setState(157);
+			setState(162);
 			match(If);
-			setState(158);
+			setState(163);
 			((If_expressionContext)_localctx).cond = expression(0);
-			setState(177);
+			setState(182);
 			_errHandler.sync(this);
 			switch ( getInterpreter().adaptivePredict(_input,9,_ctx) ) {
 			case 1:
 				{
 				{
-				setState(159);
+				setState(164);
 				match(ExitWith);
-				setState(160);
+				setState(165);
 				((If_expressionContext)_localctx).exitWith = expression(0);
 				((If_expressionContext)_localctx).ast =  new AST.IfExpr(((If_expressionContext)_localctx).cond.ast, ((If_expressionContext)_localctx).exitWith.ast, null, AST.IfExpr.Type.ExitWith);
 				}
@@ -900,9 +936,9 @@ public class ExpressionParser extends Parser {
 			case 2:
 				{
 				{
-				setState(163);
+				setState(168);
 				match(Then);
-				setState(164);
+				setState(169);
 				((If_expressionContext)_localctx).arr = array();
 				((If_expressionContext)_localctx).ast =  new AST.IfExpr(((If_expressionContext)_localctx).cond.ast, ((If_expressionContext)_localctx).arr.ast);
 				}
@@ -911,13 +947,13 @@ public class ExpressionParser extends Parser {
 			case 3:
 				{
 				{
-				setState(167);
+				setState(172);
 				match(Then);
-				setState(168);
+				setState(173);
 				((If_expressionContext)_localctx).condIsTrue = expression(0);
-				setState(169);
+				setState(174);
 				match(Else);
-				setState(170);
+				setState(175);
 				((If_expressionContext)_localctx).condIsFalse = expression(0);
 				((If_expressionContext)_localctx).ast =  new AST.IfExpr(((If_expressionContext)_localctx).cond.ast, ((If_expressionContext)_localctx).condIsTrue.ast, ((If_expressionContext)_localctx).condIsFalse.ast, AST.IfExpr.Type.IfThen);
 				}
@@ -926,9 +962,9 @@ public class ExpressionParser extends Parser {
 			case 4:
 				{
 				{
-				setState(173);
+				setState(178);
 				match(Then);
-				setState(174);
+				setState(179);
 				((If_expressionContext)_localctx).condIsTrue = expression(0);
 				((If_expressionContext)_localctx).ast =  new AST.IfExpr(((If_expressionContext)_localctx).cond.ast, ((If_expressionContext)_localctx).condIsTrue.ast, null, AST.IfExpr.Type.IfThen);
 				}
@@ -987,29 +1023,29 @@ public class ExpressionParser extends Parser {
 		try {
 			enterOuterAlt(_localctx, 1);
 			{
-			setState(179);
+			setState(184);
 			match(LBracket);
-			setState(180);
+			setState(185);
 			((ArrayContext)_localctx).e1 = expression(0);
 			_localctx.items.add(((ArrayContext)_localctx).e1.ast);
-			setState(188);
+			setState(193);
 			_errHandler.sync(this);
 			_la = _input.LA(1);
 			while (_la==Comma) {
 				{
 				{
-				setState(182);
+				setState(187);
 				match(Comma);
-				setState(183);
+				setState(188);
 				((ArrayContext)_localctx).e2 = expression(0);
 				_localctx.items.add(((ArrayContext)_localctx).e2.ast);
 				}
 				}
-				setState(190);
+				setState(195);
 				_errHandler.sync(this);
 				_la = _input.LA(1);
 			}
-			setState(191);
+			setState(196);
 			match(RBracket);
 			((ArrayContext)_localctx).ast =  new AST.Array(_localctx.items);
 			}
@@ -1049,13 +1085,13 @@ public class ExpressionParser extends Parser {
 		Int_valueContext _localctx = new Int_valueContext(_ctx, getState());
 		enterRule(_localctx, 20, RULE_int_value);
 		try {
-			setState(198);
+			setState(203);
 			_errHandler.sync(this);
 			switch (_input.LA(1)) {
 			case IntegerLiteral:
 				enterOuterAlt(_localctx, 1);
 				{
-				setState(194);
+				setState(199);
 				((Int_valueContext)_localctx).il = match(IntegerLiteral);
 				((Int_valueContext)_localctx).i =  new Integer((((Int_valueContext)_localctx).il!=null?((Int_valueContext)_localctx).il.getText():null));
 				}
@@ -1063,7 +1099,7 @@ public class ExpressionParser extends Parser {
 			case HexLiteral:
 				enterOuterAlt(_localctx, 2);
 				{
-				setState(196);
+				setState(201);
 				((Int_valueContext)_localctx).hl = match(HexLiteral);
 				((Int_valueContext)_localctx).i =  new Integer(Integer.decode((((Int_valueContext)_localctx).hl!=null?((Int_valueContext)_localctx).hl.getText():null)));
 				}
@@ -1107,7 +1143,7 @@ public class ExpressionParser extends Parser {
 		try {
 			enterOuterAlt(_localctx, 1);
 			{
-			setState(200);
+			setState(205);
 			((Float_valueContext)_localctx).fl = match(FloatLiteral);
 			((Float_valueContext)_localctx).d =  new Double((((Float_valueContext)_localctx).fl!=null?((Float_valueContext)_localctx).fl.getText():null));
 			}
@@ -1133,85 +1169,89 @@ public class ExpressionParser extends Parser {
 	private boolean expression_sempred(ExpressionContext _localctx, int predIndex) {
 		switch (predIndex) {
 		case 0:
-			return precpred(_ctx, 10);
+			return precpred(_ctx, 11);
 		case 1:
-			return precpred(_ctx, 9);
+			return precpred(_ctx, 10);
 		case 2:
-			return precpred(_ctx, 8);
+			return precpred(_ctx, 9);
 		case 3:
-			return precpred(_ctx, 7);
+			return precpred(_ctx, 8);
 		case 4:
-			return precpred(_ctx, 5);
+			return precpred(_ctx, 7);
 		case 5:
-			return precpred(_ctx, 4);
+			return precpred(_ctx, 5);
 		case 6:
+			return precpred(_ctx, 4);
+		case 7:
 			return precpred(_ctx, 3);
 		}
 		return true;
 	}
 
 	public static final String _serializedATN =
-		"\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964\3$\u00ce\4\2\t\2\4"+
+		"\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964\3*\u00d3\4\2\t\2\4"+
 		"\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t\7\4\b\t\b\4\t\t\t\4\n\t\n\4\13\t"+
 		"\13\4\f\t\f\4\r\t\r\3\2\3\2\3\2\3\2\7\2\37\n\2\f\2\16\2\"\13\2\3\2\3\2"+
 		"\5\2&\n\2\3\2\3\2\3\3\3\3\3\3\3\3\3\3\3\3\5\3\60\n\3\3\4\3\4\3\4\3\4\3"+
 		"\4\3\5\3\5\3\5\3\5\5\5;\n\5\3\5\3\5\3\5\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3"+
 		"\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\5\6P\n\6\3\6\3\6\3\6\3\6\3\6\3\6\3"+
 		"\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6"+
-		"\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\7\6u\n\6\f\6\16\6x\13\6\3"+
-		"\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\5\7\u008a"+
-		"\n\7\3\b\3\b\3\b\3\b\3\b\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\3"+
-		"\t\3\t\5\t\u009e\n\t\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3"+
-		"\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\5\n\u00b4\n\n\3\13\3\13\3\13\3\13\3\13"+
-		"\3\13\3\13\7\13\u00bd\n\13\f\13\16\13\u00c0\13\13\3\13\3\13\3\13\3\f\3"+
-		"\f\3\f\3\f\5\f\u00c9\n\f\3\r\3\r\3\r\3\r\2\3\n\16\2\4\6\b\n\f\16\20\22"+
-		"\24\26\30\2\2\2\u00dc\2 \3\2\2\2\4/\3\2\2\2\6\61\3\2\2\2\b\66\3\2\2\2"+
-		"\nO\3\2\2\2\f\u0089\3\2\2\2\16\u008b\3\2\2\2\20\u009d\3\2\2\2\22\u009f"+
-		"\3\2\2\2\24\u00b5\3\2\2\2\26\u00c8\3\2\2\2\30\u00ca\3\2\2\2\32\33\5\4"+
-		"\3\2\33\34\7\31\2\2\34\35\b\2\1\2\35\37\3\2\2\2\36\32\3\2\2\2\37\"\3\2"+
-		"\2\2 \36\3\2\2\2 !\3\2\2\2!#\3\2\2\2\" \3\2\2\2#%\5\4\3\2$&\7\31\2\2%"+
-		"$\3\2\2\2%&\3\2\2\2&\'\3\2\2\2\'(\b\2\1\2(\3\3\2\2\2)*\5\6\4\2*+\b\3\1"+
-		"\2+\60\3\2\2\2,-\5\n\6\2-.\b\3\1\2.\60\3\2\2\2/)\3\2\2\2/,\3\2\2\2\60"+
-		"\5\3\2\2\2\61\62\7\32\2\2\62\63\7\30\2\2\63\64\5\n\6\2\64\65\b\4\1\2\65"+
-		"\7\3\2\2\2\66:\7\6\2\2\678\5\2\2\289\b\5\1\29;\3\2\2\2:\67\3\2\2\2:;\3"+
-		"\2\2\2;<\3\2\2\2<=\7\7\2\2=>\b\5\1\2>\t\3\2\2\2?@\b\6\1\2@A\5\f\7\2AB"+
-		"\b\6\1\2BP\3\2\2\2CD\5\16\b\2DE\b\6\1\2EP\3\2\2\2FG\5\20\t\2GH\b\6\1\2"+
-		"HP\3\2\2\2IJ\5\22\n\2JK\b\6\1\2KP\3\2\2\2LM\5\b\5\2MN\b\6\1\2NP\3\2\2"+
-		"\2O?\3\2\2\2OC\3\2\2\2OF\3\2\2\2OI\3\2\2\2OL\3\2\2\2Pv\3\2\2\2QR\f\f\2"+
-		"\2RS\7\r\2\2ST\5\n\6\rTU\b\6\1\2Uu\3\2\2\2VW\f\13\2\2WX\7\f\2\2XY\5\n"+
-		"\6\fYZ\b\6\1\2Zu\3\2\2\2[\\\f\n\2\2\\]\7\n\2\2]^\5\n\6\13^_\b\6\1\2_u"+
-		"\3\2\2\2`a\f\t\2\2ab\7\13\2\2bc\5\n\6\ncd\b\6\1\2du\3\2\2\2ef\f\7\2\2"+
-		"fg\7\22\2\2gh\5\n\6\bhi\b\6\1\2iu\3\2\2\2jk\f\6\2\2kl\7\21\2\2lm\5\n\6"+
-		"\7mn\b\6\1\2nu\3\2\2\2op\f\5\2\2pq\7\27\2\2qr\5\n\6\6rs\b\6\1\2su\3\2"+
-		"\2\2tQ\3\2\2\2tV\3\2\2\2t[\3\2\2\2t`\3\2\2\2te\3\2\2\2tj\3\2\2\2to\3\2"+
-		"\2\2ux\3\2\2\2vt\3\2\2\2vw\3\2\2\2w\13\3\2\2\2xv\3\2\2\2yz\7\n\2\2z{\5"+
-		"\16\b\2{|\b\7\1\2|\u008a\3\2\2\2}~\7\n\2\2~\177\5\20\t\2\177\u0080\b\7"+
-		"\1\2\u0080\u008a\3\2\2\2\u0081\u0082\7\13\2\2\u0082\u0083\5\16\b\2\u0083"+
-		"\u0084\b\7\1\2\u0084\u008a\3\2\2\2\u0085\u0086\7\13\2\2\u0086\u0087\5"+
-		"\20\t\2\u0087\u0088\b\7\1\2\u0088\u008a\3\2\2\2\u0089y\3\2\2\2\u0089}"+
-		"\3\2\2\2\u0089\u0081\3\2\2\2\u0089\u0085\3\2\2\2\u008a\r\3\2\2\2\u008b"+
-		"\u008c\7\16\2\2\u008c\u008d\5\n\6\2\u008d\u008e\7\17\2\2\u008e\u008f\b"+
-		"\b\1\2\u008f\17\3\2\2\2\u0090\u0091\7\32\2\2\u0091\u009e\b\t\1\2\u0092"+
-		"\u0093\5\26\f\2\u0093\u0094\b\t\1\2\u0094\u009e\3\2\2\2\u0095\u0096\5"+
-		"\30\r\2\u0096\u0097\b\t\1\2\u0097\u009e\3\2\2\2\u0098\u0099\7\3\2\2\u0099"+
-		"\u009e\b\t\1\2\u009a\u009b\5\24\13\2\u009b\u009c\b\t\1\2\u009c\u009e\3"+
-		"\2\2\2\u009d\u0090\3\2\2\2\u009d\u0092\3\2\2\2\u009d\u0095\3\2\2\2\u009d"+
-		"\u0098\3\2\2\2\u009d\u009a\3\2\2\2\u009e\21\3\2\2\2\u009f\u00a0\7\23\2"+
-		"\2\u00a0\u00b3\5\n\6\2\u00a1\u00a2\7\26\2\2\u00a2\u00a3\5\n\6\2\u00a3"+
-		"\u00a4\b\n\1\2\u00a4\u00b4\3\2\2\2\u00a5\u00a6\7\24\2\2\u00a6\u00a7\5"+
-		"\24\13\2\u00a7\u00a8\b\n\1\2\u00a8\u00b4\3\2\2\2\u00a9\u00aa\7\24\2\2"+
-		"\u00aa\u00ab\5\n\6\2\u00ab\u00ac\7\25\2\2\u00ac\u00ad\5\n\6\2\u00ad\u00ae"+
-		"\b\n\1\2\u00ae\u00b4\3\2\2\2\u00af\u00b0\7\24\2\2\u00b0\u00b1\5\n\6\2"+
-		"\u00b1\u00b2\b\n\1\2\u00b2\u00b4\3\2\2\2\u00b3\u00a1\3\2\2\2\u00b3\u00a5"+
-		"\3\2\2\2\u00b3\u00a9\3\2\2\2\u00b3\u00af\3\2\2\2\u00b4\23\3\2\2\2\u00b5"+
-		"\u00b6\7\b\2\2\u00b6\u00b7\5\n\6\2\u00b7\u00be\b\13\1\2\u00b8\u00b9\7"+
-		"\20\2\2\u00b9\u00ba\5\n\6\2\u00ba\u00bb\b\13\1\2\u00bb\u00bd\3\2\2\2\u00bc"+
-		"\u00b8\3\2\2\2\u00bd\u00c0\3\2\2\2\u00be\u00bc\3\2\2\2\u00be\u00bf\3\2"+
-		"\2\2\u00bf\u00c1\3\2\2\2\u00c0\u00be\3\2\2\2\u00c1\u00c2\7\t\2\2\u00c2"+
-		"\u00c3\b\13\1\2\u00c3\25\3\2\2\2\u00c4\u00c5\7\33\2\2\u00c5\u00c9\b\f"+
-		"\1\2\u00c6\u00c7\7 \2\2\u00c7\u00c9\b\f\1\2\u00c8\u00c4\3\2\2\2\u00c8"+
-		"\u00c6\3\2\2\2\u00c9\27\3\2\2\2\u00ca\u00cb\7\34\2\2\u00cb\u00cc\b\r\1"+
-		"\2\u00cc\31\3\2\2\2\16 %/:Otv\u0089\u009d\u00b3\u00be\u00c8";
+		"\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\7\6z"+
+		"\n\6\f\6\16\6}\13\6\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3"+
+		"\7\3\7\3\7\3\7\5\7\u008f\n\7\3\b\3\b\3\b\3\b\3\b\3\t\3\t\3\t\3\t\3\t\3"+
+		"\t\3\t\3\t\3\t\3\t\3\t\3\t\3\t\5\t\u00a3\n\t\3\n\3\n\3\n\3\n\3\n\3\n\3"+
+		"\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\3\n\5\n\u00b9\n\n\3"+
+		"\13\3\13\3\13\3\13\3\13\3\13\3\13\7\13\u00c2\n\13\f\13\16\13\u00c5\13"+
+		"\13\3\13\3\13\3\13\3\f\3\f\3\f\3\f\5\f\u00ce\n\f\3\r\3\r\3\r\3\r\2\3\n"+
+		"\16\2\4\6\b\n\f\16\20\22\24\26\30\2\3\3\2\30\35\2\u00e2\2 \3\2\2\2\4/"+
+		"\3\2\2\2\6\61\3\2\2\2\b\66\3\2\2\2\nO\3\2\2\2\f\u008e\3\2\2\2\16\u0090"+
+		"\3\2\2\2\20\u00a2\3\2\2\2\22\u00a4\3\2\2\2\24\u00ba\3\2\2\2\26\u00cd\3"+
+		"\2\2\2\30\u00cf\3\2\2\2\32\33\5\4\3\2\33\34\7\37\2\2\34\35\b\2\1\2\35"+
+		"\37\3\2\2\2\36\32\3\2\2\2\37\"\3\2\2\2 \36\3\2\2\2 !\3\2\2\2!#\3\2\2\2"+
+		"\" \3\2\2\2#%\5\4\3\2$&\7\37\2\2%$\3\2\2\2%&\3\2\2\2&\'\3\2\2\2\'(\b\2"+
+		"\1\2(\3\3\2\2\2)*\5\6\4\2*+\b\3\1\2+\60\3\2\2\2,-\5\n\6\2-.\b\3\1\2.\60"+
+		"\3\2\2\2/)\3\2\2\2/,\3\2\2\2\60\5\3\2\2\2\61\62\7 \2\2\62\63\7\36\2\2"+
+		"\63\64\5\n\6\2\64\65\b\4\1\2\65\7\3\2\2\2\66:\7\6\2\2\678\5\2\2\289\b"+
+		"\5\1\29;\3\2\2\2:\67\3\2\2\2:;\3\2\2\2;<\3\2\2\2<=\7\7\2\2=>\b\5\1\2>"+
+		"\t\3\2\2\2?@\b\6\1\2@A\5\f\7\2AB\b\6\1\2BP\3\2\2\2CD\5\16\b\2DE\b\6\1"+
+		"\2EP\3\2\2\2FG\5\20\t\2GH\b\6\1\2HP\3\2\2\2IJ\5\22\n\2JK\b\6\1\2KP\3\2"+
+		"\2\2LM\5\b\5\2MN\b\6\1\2NP\3\2\2\2O?\3\2\2\2OC\3\2\2\2OF\3\2\2\2OI\3\2"+
+		"\2\2OL\3\2\2\2P{\3\2\2\2QR\f\r\2\2RS\7\r\2\2ST\5\n\6\16TU\b\6\1\2Uz\3"+
+		"\2\2\2VW\f\f\2\2WX\7\f\2\2XY\5\n\6\rYZ\b\6\1\2Zz\3\2\2\2[\\\f\13\2\2\\"+
+		"]\7\n\2\2]^\5\n\6\f^_\b\6\1\2_z\3\2\2\2`a\f\n\2\2ab\7\13\2\2bc\5\n\6\13"+
+		"cd\b\6\1\2dz\3\2\2\2ef\f\t\2\2fg\t\2\2\2gh\5\n\6\nhi\b\6\1\2iz\3\2\2\2"+
+		"jk\f\7\2\2kl\7\22\2\2lm\5\n\6\bmn\b\6\1\2nz\3\2\2\2op\f\6\2\2pq\7\21\2"+
+		"\2qr\5\n\6\7rs\b\6\1\2sz\3\2\2\2tu\f\5\2\2uv\7\27\2\2vw\5\n\6\6wx\b\6"+
+		"\1\2xz\3\2\2\2yQ\3\2\2\2yV\3\2\2\2y[\3\2\2\2y`\3\2\2\2ye\3\2\2\2yj\3\2"+
+		"\2\2yo\3\2\2\2yt\3\2\2\2z}\3\2\2\2{y\3\2\2\2{|\3\2\2\2|\13\3\2\2\2}{\3"+
+		"\2\2\2~\177\7\n\2\2\177\u0080\5\16\b\2\u0080\u0081\b\7\1\2\u0081\u008f"+
+		"\3\2\2\2\u0082\u0083\7\n\2\2\u0083\u0084\5\20\t\2\u0084\u0085\b\7\1\2"+
+		"\u0085\u008f\3\2\2\2\u0086\u0087\7\13\2\2\u0087\u0088\5\16\b\2\u0088\u0089"+
+		"\b\7\1\2\u0089\u008f\3\2\2\2\u008a\u008b\7\13\2\2\u008b\u008c\5\20\t\2"+
+		"\u008c\u008d\b\7\1\2\u008d\u008f\3\2\2\2\u008e~\3\2\2\2\u008e\u0082\3"+
+		"\2\2\2\u008e\u0086\3\2\2\2\u008e\u008a\3\2\2\2\u008f\r\3\2\2\2\u0090\u0091"+
+		"\7\16\2\2\u0091\u0092\5\n\6\2\u0092\u0093\7\17\2\2\u0093\u0094\b\b\1\2"+
+		"\u0094\17\3\2\2\2\u0095\u0096\7 \2\2\u0096\u00a3\b\t\1\2\u0097\u0098\5"+
+		"\26\f\2\u0098\u0099\b\t\1\2\u0099\u00a3\3\2\2\2\u009a\u009b\5\30\r\2\u009b"+
+		"\u009c\b\t\1\2\u009c\u00a3\3\2\2\2\u009d\u009e\7\3\2\2\u009e\u00a3\b\t"+
+		"\1\2\u009f\u00a0\5\24\13\2\u00a0\u00a1\b\t\1\2\u00a1\u00a3\3\2\2\2\u00a2"+
+		"\u0095\3\2\2\2\u00a2\u0097\3\2\2\2\u00a2\u009a\3\2\2\2\u00a2\u009d\3\2"+
+		"\2\2\u00a2\u009f\3\2\2\2\u00a3\21\3\2\2\2\u00a4\u00a5\7\23\2\2\u00a5\u00b8"+
+		"\5\n\6\2\u00a6\u00a7\7\26\2\2\u00a7\u00a8\5\n\6\2\u00a8\u00a9\b\n\1\2"+
+		"\u00a9\u00b9\3\2\2\2\u00aa\u00ab\7\24\2\2\u00ab\u00ac\5\24\13\2\u00ac"+
+		"\u00ad\b\n\1\2\u00ad\u00b9\3\2\2\2\u00ae\u00af\7\24\2\2\u00af\u00b0\5"+
+		"\n\6\2\u00b0\u00b1\7\25\2\2\u00b1\u00b2\5\n\6\2\u00b2\u00b3\b\n\1\2\u00b3"+
+		"\u00b9\3\2\2\2\u00b4\u00b5\7\24\2\2\u00b5\u00b6\5\n\6\2\u00b6\u00b7\b"+
+		"\n\1\2\u00b7\u00b9\3\2\2\2\u00b8\u00a6\3\2\2\2\u00b8\u00aa\3\2\2\2\u00b8"+
+		"\u00ae\3\2\2\2\u00b8\u00b4\3\2\2\2\u00b9\23\3\2\2\2\u00ba\u00bb\7\b\2"+
+		"\2\u00bb\u00bc\5\n\6\2\u00bc\u00c3\b\13\1\2\u00bd\u00be\7\20\2\2\u00be"+
+		"\u00bf\5\n\6\2\u00bf\u00c0\b\13\1\2\u00c0\u00c2\3\2\2\2\u00c1\u00bd\3"+
+		"\2\2\2\u00c2\u00c5\3\2\2\2\u00c3\u00c1\3\2\2\2\u00c3\u00c4\3\2\2\2\u00c4"+
+		"\u00c6\3\2\2\2\u00c5\u00c3\3\2\2\2\u00c6\u00c7\7\t\2\2\u00c7\u00c8\b\13"+
+		"\1\2\u00c8\25\3\2\2\2\u00c9\u00ca\7!\2\2\u00ca\u00ce\b\f\1\2\u00cb\u00cc"+
+		"\7&\2\2\u00cc\u00ce\b\f\1\2\u00cd\u00c9\3\2\2\2\u00cd\u00cb\3\2\2\2\u00ce"+
+		"\27\3\2\2\2\u00cf\u00d0\7\"\2\2\u00d0\u00d1\b\r\1\2\u00d1\31\3\2\2\2\16"+
+		" %/:Oy{\u008e\u00a2\u00b8\u00c3\u00cd";
 	public static final ATN _ATN =
 		new ATNDeserializer().deserialize(_serializedATN.toCharArray());
 	static {